requirements.txt
aniso8601==9.0.1
apispec==5.1.1
apispec-webframeworks==0.5.2
appnope==0.1.2
astroid==2.8.4
asttokens==2.0.5
backcall==0.2.0
backports.entry-points-selectable==1.1.1
black==22.1.0
bloom-pick @ file:///Users/manuel/Downloads/apps/bloomscape/bloompick-core
boto3==1.19.3
botocore==1.22.3
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.3
coverage==6.0.2
cryptography==35.0.0
decorator==5.1.1
distlib==0.3.3
dodgy==0.2.1
elastic-transport==8.1.1
elasticsearch==8.1.2
executing==0.8.2
filelock==3.4.0
flake8==3.9.2
flake8-polyfill==1.0.2
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Limiter==1.4
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
gunicorn==20.1.0
idna==2.10
ipython==8.0.1
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.2
jmespath==0.10.0
lazy-object-proxy==1.6.0
limits==2.3.3
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
newrelic==7.2.2.169
numpy==1.21.3
oauthlib==3.2.0
pandas==1.3.4
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pep8-naming==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.11.15
platformdirs==2.4.0
probableparsing==0.0.1
prompt-toolkit==3.0.28
prospector==1.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.11.2
pylint==2.11.1
pylint-celery==0.3
pylint-django==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
PyMySQL==1.0.2
PyPDF2==1.26.0
python-barcode==0.13.1
python-crfsuite==0.9.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.24.0
requests-oauthlib==1.3.0
requirements-detector==0.7
s3transfer==0.5.0
setoptconf-tmp==0.3.1
shippo==2.0.2
simplejson==3.17.2
six==1.16.0
snowballstemmer==2.1.0
SQLAlchemy==1.4.26
stack-data==0.1.4
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.9
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webargs==8.0.1
Werkzeug==2.0.2
wrapt==1.13.2
